+++ src/test/auxiliary/namespaced_enums.rs
@@ -10,12 +10,11 @@
 
 pub enum Foo {
     A,
-    B(int),
-    C { a: int },
+    B(isize),
+    C { a: isize },
 }
 
 impl Foo {
     pub fn foo() {}
     pub fn bar(&self) {}
 }
-
