+++ src/test/run-make/extern-fn-reachable/main.rs
@@ -8,17 +8,20 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+#![feature(std_misc, old_path, os)]
+
 use std::dynamic_lib::DynamicLibrary;
 use std::os;
+use std::path::Path;
 
 pub fn main() {
     unsafe {
         let path = Path::new("libdylib.so");
         let a = DynamicLibrary::open(Some(&path)).unwrap();
-        assert!(a.symbol::<int>("fun1").is_ok());
-        assert!(a.symbol::<int>("fun2").is_err());
-        assert!(a.symbol::<int>("fun3").is_err());
-        assert!(a.symbol::<int>("fun4").is_ok());
-        assert!(a.symbol::<int>("fun5").is_ok());
+        assert!(a.symbol::<isize>("fun1").is_ok());
+        assert!(a.symbol::<isize>("fun2").is_err());
+        assert!(a.symbol::<isize>("fun3").is_err());
+        assert!(a.symbol::<isize>("fun4").is_ok());
+        assert!(a.symbol::<isize>("fun5").is_ok());
     }
 }
