+++ src/test/run-pass/associated-types-basic.rs
@@ -8,6 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+// pretty-expanded FIXME #23616
+
+#![feature(core)]
+
 use std::marker::MarkerTrait;
 
 trait Foo : MarkerTrait {
@@ -15,11 +19,11 @@ trait Foo : MarkerTrait {
 }
 
 impl Foo for i32 {
-    type T = int;
+    type T = isize;
 }
 
 fn main() {
     let x: <i32 as Foo>::T = 22;
-    let y: int = 44;
+    let y: isize = 44;
     assert_eq!(x * 2, y);
 }
