+++ src/test/run-pass/associated-types-ref-in-struct-literal.rs
@@ -10,14 +10,16 @@
 
 // Test associated type references in a struct literal. Issue #20535.
 
+// pretty-expanded FIXME #23616
+
 pub trait Foo {
     type Bar;
 
     fn dummy(&self) { }
 }
 
-impl Foo for int {
-    type Bar = int;
+impl Foo for isize {
+    type Bar = isize;
 }
 
 struct Thing<F: Foo> {
