+++ src/test/run-pass/borrowck-scope-of-deref-issue-4666.rs
@@ -12,15 +12,17 @@
 // limited to the deref operation itself, and does not infect the
 // block as a whole.
 
+// pretty-expanded FIXME #23616
+
 struct Box {
-    x: uint
+    x: usize
 }
 
 impl Box {
-    fn get(&self) -> &uint {
+    fn get(&self) -> &usize {
         &self.x
     }
-    fn set(&mut self, x: uint) {
+    fn set(&mut self, x: usize) {
         self.x = x;
     }
 }
