+++ src/test/run-pass/bug-7183-generics.rs
@@ -8,6 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+// pretty-expanded FIXME #23616
+
 trait Speak : Sized {
     fn say(&self, s:&str) -> String;
     fn hi(&self) -> String { hello(self) }
@@ -17,7 +19,7 @@ fn hello<S:Speak>(s:&S) -> String{
     s.say("hello")
 }
 
-impl Speak for int {
+impl Speak for isize {
     fn say(&self, s:&str) -> String {
         format!("{}: {}", s, *self)
     }
@@ -37,8 +39,8 @@ pub fn main() {
     assert_eq!(3.hi(), "hello: 3".to_string());
     assert_eq!(Some(Some(3)).hi(),
                "something!something!hello: 3".to_string());
-    assert_eq!(None::<int>.hi(), "hello - none".to_string());
+    assert_eq!(None::<isize>.hi(), "hello - none".to_string());
 
-    assert_eq!(Some(None::<int>).hi(), "something!hello - none".to_string());
+    assert_eq!(Some(None::<isize>).hi(), "something!hello - none".to_string());
     assert_eq!(Some(3).hi(), "something!hello: 3".to_string());
 }
